requirements.txt
tensorflow
numpy
matplotlib
redis
flask
flask-sqlalchemy
psycopg2
requests
python-dotenv
